6126928844029890469
ManifestID: 6126928844029890469
Date: 2026-06-28T23:44:27+08:00
Version: TB-246879

+++ Assembly-CSharp/MeterScreen.cs
@@ -121,7 +121,7 @@ public class MeterScreen : KScreen
 	{
 		if (this.RationsText != null && RationTracker.Get() != null)
 		{
-			int num = (int)RationTracker.Get().CountRations(null, true);
+			long num = (long)RationTracker.Get().CountRations(null, true);
 			this.RationsText.text = GameUtil.GetFormattedCalories((float)num, GameUtil.TimeSlice.None, true);
 		}
 	}
